requirements.txt
google-adk>=0.3.0
openai>=1.0.0
nemoguardrails>=0.10.0
streamlit>=1.35.0
